+++ test/libsolidity/semanticTests/array/pop/byte_array_pop_long_storage_empty.sol
@@ -7,7 +7,7 @@ contract c {
         for (uint8 i = 0; i <= 40; i++)
             data.push(byte(i+1));
         for (int8 j = 40; j >= 0; j--) {
-            require(data[uint8(j)] == byte(j+1));
+            require(data[uint8(j)] == byte(uint8(j+1)));
             require(data.length == uint8(j+1));
             data.pop();
         }
