+++ test/libsolidity/semanticTests/storage/packed_storage_signed.sol
@@ -13,9 +13,9 @@ contract C {
             b = (0 - uint8(a)) * 2;
             c = a * int8(120) * int8(121);
         }
-        x1 = uint256(a);
+        x1 = uint256(int256(a));
         x2 = b;
-        x3 = uint256(c);
+        x3 = uint256(int256(c));
         x4 = d;
     }
 }
